Show all subcommands in top-level help
The `run` function was not correctly generating the help page for the
top-level command when subcommands were present. The `augmentedParser`
was too complex for `getDocPage` to correctly extract the list of
subcommands.

This change introduces a `helpGeneratorParser` that is created on-the-fly
by combining the user's parser with the `help` and `version` command
parsers. This simpler structure allows `getDocPage` to correctly
generate the documentation for the top-level command, including all
user-defined subcommands as well as the `help` and `version` commands.