Weaken warning in READMEs
The previous [!CAUTION] block in the README files for the root package and
all sub-packages was overly strong, potentially discouraging new users.

This commit replaces the caution with a more moderate [!WARNING] that still
informs users about the API's beta status and the possibility of breaking
changes before the 1.0 release, but in a less alarming tone.

[ci skip]